+++ src/hipo/interpreter.cljs
@@ -8,69 +8,39 @@
 (def +svg-ns+ "http://www.w3.org/2000/svg")
 (def +svg-tags+ #{"svg" "g" "rect" "circle" "clipPath" "path" "line" "polygon" "polyline" "text" "textPath"})
 
-(defprotocol AttributeHandler
-  (-set [_ el n ov nv]))
-
-#_
-(deftype AsPropertyAttributeHandler []
-  AttributeHandler
-  (-set [_ el n ov nv]
-    ))
-
-(deftype DefaultAttributeHandler []
-  AttributeHandler
-  (-set [_ el n _ nv]
-    (if nv
-      (.setAttribute el n nv)
-      (.removeAttribute el n))
-    true))
-
 (defn- listener-name? [s] (identical? 0 (.indexOf s "on-")))
 (defn- listener-name->event-name [s] (.substring s 3))
 
-(deftype ListenerAttributeHandler []
-  AttributeHandler
-  (-set [_ el n ov nv]
-    (when (listener-name? n)
+(defn set-attribute!
+  [el n ov nv]
+  (if (listener-name? n)
+    (do
       (if ov
         (.removeEventListener el (listener-name->event-name n) ov))
       (if nv
-        (.addEventListener el (listener-name->event-name n) nv))
-      true)))
-
-(def ^:dynamic *default-attribute-handlers* [(ListenerAttributeHandler.) (DefaultAttributeHandler.)])
-
-(defn attribute-handlers
-  [m]
-  (if-let [s (:attribute-handlers m)]
-    (apply conj s *default-attribute-handlers*)
-    *default-attribute-handlers*))
-
-(defn set-attribute!
-  [el n ov nv ahs]
-  (loop [s ahs]
-    (if-let [f (first s)]
-      (if-not (-set f el n ov nv)
-        (recur (rest s))))))
+        (.addEventListener el (listener-name->event-name n) nv)))
+    (if nv
+      (.setAttribute el n nv)
+      (.removeAttribute el n))))
 
 (declare create-child)
 
 (defn append-child!
-  [el o ahs]
-  (.appendChild el (create-child o ahs)))
+  [el o]
+  (.appendChild el (create-child o)))
 
 (defn append-children!
-  [el o ahs]
+  [el o]
   (if (seq? o)
     (loop [s o]
       (when-not (empty? s)
         (if-let [h (first s)]
-          (append-children! el h ahs))
+          (append-children! el h))
         (recur (rest s))))
-    (append-child! el o ahs)))
+    (append-child! el o)))
 
 (defn create-vector
-  [[node-key & rest] ahs]
+  [[node-key & rest]]
   (let [literal-attrs (when-let [f (first rest)] (when (map? f) f))
         children (if literal-attrs (drop 1 rest) rest)
         node (name node-key)
@@ -86,9 +56,9 @@
         (set! (.-id el) id))
       (doseq [[k v] (dissoc literal-attrs :class)]
         (if v
-          (set-attribute! el (name k) nil v ahs)))
+          (set-attribute! el (name k) nil v)))
       (if (seq children)
-        (append-children! el children ahs))
+        (append-children! el children))
       el)))
 
 (defn mark-as-partially-compiled!
@@ -98,44 +68,44 @@
     (do (aset el "hipo-partially-compiled" true) el)))
 
 (defn create-child
-  [o ahs]
+  [o]
   {:pre [(or (hic/literal? o) (vector? o))]}
   (if (hic/literal? o) ; literal check is much more efficient than vector check
     (.createTextNode js/document o)
-    (create-vector o ahs)))
+    (create-vector o)))
 
 (defn append-to-parent
-  [el o ahs]
+  [el o]
   {:pre [(not (nil? o))]}
   (mark-as-partially-compiled! el)
-  (append-children! el o ahs))
+  (append-children! el o))
 
 (defn create
-  [o ahs]
+  [o]
   {:pre [(not (nil? o))]}
   (mark-as-partially-compiled!
     (if (seq? o)
       (let [f (.createDocumentFragment js/document)]
-        (append-children! f o ahs)
+        (append-children! f o)
         f)
-      (create-child o ahs))))
+      (create-child o))))
 
 ; Update
 
 (defn update-attributes!
-  [el om nm int ahs]
+  [el om nm int]
   (doseq [[nk nv] nm
           :let [ov (nk om) n (name nk)]]
     (if-not (identical? ov nv)
       (if nv
         (intercept int :update-attribute {:target el :name n :value nv}
-          (set-attribute! el n ov nv ahs))
+          (set-attribute! el n ov nv))
         (intercept int :remove-attribute {:target el :name n}
-          (set-attribute! el n ov nil ahs)))))
+          (set-attribute! el n ov nil)))))
   (doseq [k (set/difference (set (keys om)) (set (keys nm)))
           :let [n (name k) ov (k om)]]
     (intercept int :remove-attribute {:target el :name n}
-      (set-attribute! el n ov nil ahs))))
+      (set-attribute! el n ov nil))))
 
 (declare update!)
 
@@ -144,7 +114,7 @@
 (defn keyed-children->indexed-map [v] (into {} (for [ih (map-indexed (fn [idx itm] [idx itm]) v)] [(child-key (nth ih 1)) ih])))
 
 (defn update-keyed-children!
-  [el och nch int ahs]
+  [el och nch int]
   (let [om (keyed-children->map och)
         nm (keyed-children->indexed-map nch)
         cs (dom/children el (apply max (set/intersection (set (keys nm)) (set (keys om)))))]
@@ -155,10 +125,10 @@
           (if (identical? oh h)
             (dom/insert-child-at! el ii (nth cs i))
             (let [ncel (.removeChild el (nth cs i))]
-              (update! ncel oh h int ahs)
+              (update! ncel oh h int)
               (dom/insert-child-at! el ii ncel)))); TODO improve perf by relying on (cs ii)? index should be updated based on new insertions
         ; new node
-        (let [nel (create-child h ahs)]
+        (let [nel (create-child h)]
           (intercept int :insert-at {:target el :value nel :index ii}
             (dom/insert-child-at! el ii nel)))))
     (let [d (count (set/difference (set (keys om)) (set (keys nm))))]
@@ -166,7 +136,7 @@
         (dom/remove-trailing-children! el d)))))
 
 (defn update-non-keyed-children!
-  [el och nch int ahs]
+  [el och nch int]
   (let [oc (count och)
         nc (count nch)
         d (- oc nc)]
@@ -181,36 +151,36 @@
             nv (nth nch i)]
         (if-not (identical? ov nv)
           (if-let [cel (dom/child-at el i)]
-            (update! cel ov nv int ahs)))))
+            (update! cel ov nv int)))))
     ; Create new elements if (count nch) > (count oh)
     (if (neg? d)
       (if (identical? -1 d)
         (let [nel (peek nch)]
           (intercept int :append {:target el :value nel}
-            (append-child! el nel ahs)))
+            (append-child! el nel)))
         (let [f (.createDocumentFragment js/document)
               cs (apply list (if (identical? 0 oc) nch (subvec nch oc)))]
           ; An intermediary DocumentFragment is used to reduce the number of append to the attached node
           (intercept int :append {:target el :value cs}
-            (append-children! f cs ahs))
+            (append-children! f cs))
           (.appendChild el f))))))
 
 (defn keyed-children? [v] (not (nil? (child-key (nth v 0)))))
 
 (defn update-children!
-  [el och nch int ahs]
+  [el och nch int]
   (if (f/emptyv? nch)
     (intercept int :clear {:target el}
       (dom/clear! el))
     (if (keyed-children? nch)
-      (update-keyed-children! el och nch int ahs)
-      (update-non-keyed-children! el och nch int ahs))))
+      (update-keyed-children! el och nch int)
+      (update-non-keyed-children! el och nch int))))
 
 (defn update-vector!
-  [el oh nh int ahs]
+  [el oh nh int]
   {:pre [(vector? oh) (vector? nh)]}
   (if-not (identical? (hic/parse-tag-name (name (nth nh 0))) (hic/parse-tag-name (name (nth oh 0))))
-    (let [nel (create nh ahs)]
+    (let [nel (create nh)]
       (intercept int :replace {:target el :value nel}
         (dom/replace! el nel)))
     (let [om (hic/attributes oh)
@@ -219,15 +189,15 @@
           nch (hic/children nh)]
       (if-not (identical? och nch)
         (intercept int :update-children {:target el}
-          (update-children! el (hic/flatten-children och) (hic/flatten-children nch) int ahs)))
+          (update-children! el (hic/flatten-children och) (hic/flatten-children nch) int)))
       (if-not (identical? om nm)
-        (update-attributes! el om nm int ahs)))))
+        (update-attributes! el om nm int)))))
 
 (defn update!
-  [el ph h int ahs]
+  [el ph h int]
   {:pre [(or (vector? h) (hic/literal? h))]}
   (if (hic/literal? h) ; literal check is much more efficient than vector check
     (if-not (identical? ph h)
       (intercept int :replace {:target el :value h}
         (dom/replace-text! el h)))
-    (update-vector! el ph h int ahs)))
+    (update-vector! el ph h int)))