+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## [0.4.0](https://github.com/jeluard/hipo/issues?q=is%3Aclosed+milestone%3A0.4.0)
 
 * **[~]** Improved Interceptor semantic
+* **[-]** Removed attribute handler extensibility
 
 ## [0.3.0](https://github.com/jeluard/hipo/issues?q=is%3Aclosed+milestone%3A0.3.0)